backend/backend-retina-analyzer/requirements.txt
numpy==1.24.3
opencv-python==4.8.0.76
scikit-image==0.21.0
scikit-learn==1.3.0
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.0
scipy==1.11.3
matplotlib==3.7.2
requests==2.31.0
pydantic==2.4.2
azure-cosmos==4.3.1
python-dotenv==1.0.0
azure-servicebus==7.11.3
azure-storage-blob==12.17.0
